Refactor CLI structure and service management
- Introduced a new CLI module with subcommands for running, installing, and uninstalling the service.
- Moved installation logic to a dedicated `install` module, handling macOS and Linux specifics.
- Created an `uninstall` module to manage service removal across platforms.
- Refactored the service running logic into a `run` module, supporting both standalone and Windows service modes.
- Simplified the main entry points for each binary to delegate to the CLI module.
- Removed redundant code and improved error handling across service management functions.
