+++ Cargo.toml
@@ -10,6 +10,11 @@ name = "clash_verge_service_ipc"
 path = "src/lib.rs"
 
 [[bin]]
+name = "clash-verge-service-cli"
+path = "src/bin/cli.rs"
+required-features = ["standalone"]
+
+[[bin]]
 name = "clash-verge-service"
 path = "src/bin/service.rs"
 required-features = ["standalone"]
@@ -45,6 +50,7 @@ standalone = [
     "response",
     "dep:platform_lib",
     "compact_str",
+    "clap",
 ]
 client = [
     "kode-bridge/client",
@@ -90,6 +96,7 @@ flexi_logger = { version = "0.31.7", optional = true }
 log = { version = "0.4.28", optional = true }
 tracing-log = { version = "0.2.0", optional = true }
 compact_str = { version = "0.9.0", features = ["serde"], optional = true }
+clap = { version = "4.5.52", features = ["derive"], optional = true }
 
 [target.'cfg(unix)'.dependencies]
 platform_lib = { package = "libc", version = "0.2.177", optional = true }
