+++ src/bin/cli.rs
@@ -1,6 +1,6 @@
 use anyhow::Error;
-use clash_verge_service_ipc::cli::run;
+use clash_verge_service_ipc::cli;
 
 fn main() -> Result<(), Error> {
-    run::main()
+    cli::main()
 }
